--- src/main/java/com/example/musing/Entity/BCategory.java
@@ -1,15 +1,0 @@
-package com.example.musing.Entity;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name="bcategory")
-public class BCategory {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
-
-    @Column(nullable = true)
-    private String categoryname;
-
-}
