--- src/main/java/com/example/musing/Entity/Prefer_Music.java
@@ -1,4 +1,0 @@
-package com.example.musing.Entity;
-
-public class Prefer_Music {
-}
