--- src/main/java/com/example/musing/Repository/Prefer_Music_Repo.java
@@ -1,4 +1,0 @@
-package com.example.musing.Repository;
-
-public class Prefer_Music_Repo {
-}
