+++ src/main/java/com/example/musing/entity/Board/Board.java
@@ -1,0 +1,65 @@
+package com.example.musing.entity.Board;
+
+import com.example.musing.entity.music.Music;
+import com.example.musing.entity.user.User;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+@Getter // Lombok 어노테이션 : 클래스 내 모든 필드의 Getter 메소드 자동 생성
+@NoArgsConstructor // Lombok 어노테이션 : 기본 생성자 자동 추가
+@Entity
+@Table(name="board")
+public class Board {
+
+    //글번호(자동증가)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+    //글 제목
+    @Column(nullable=false,length = 100)
+    private String title;
+    //글 내용
+    @Column(nullable = false,columnDefinition = "TEXT")
+    private String content;
+
+    //생성일자
+    @Column(nullable = false)
+    private LocalDateTime createdAt;
+
+    @Column(nullable = false)
+    private boolean activeCheck;
+
+    //관계설정 카테고리에 관한 외래키 보유 주인테이블
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="category", nullable = false)
+    private BCategory category;
+
+    //관계설정 유저에 관한 외래키 보유 주인테이블
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="user_id", nullable = false)
+    private User user;
+
+    //관계설정 노래에 관한 외래키 보유 주인테이블
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="music_id", nullable = false)
+    private Music music;
+
+
+    //댓글과 일대다관계 매핑
+    @OneToMany(mappedBy = "board", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Reply> replies = new ArrayList<>();
+
+    //신고내역과 일대다관계 매핑
+    @OneToMany(mappedBy = "board", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Report> Reports = new ArrayList<>();
+
+
+
+
+
+}
