+++ src/main/java/com/example/musing/entity/music/HashTag.java
@@ -1,0 +1,24 @@
+package com.example.musing.entity.music;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter // Lombok 어노테이션 : 클래스 내 모든 필드의 Getter 메소드 자동 생성
+@NoArgsConstructor // Lombok 어노테이션 : 기본 생성자 자동 추가
+@Entity
+@Table(name="hashtag")
+public class HashTag {
+
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+
+    @Column(nullable = false)
+    private String hashtag;
+
+
+
+
+}
