+++ src/main/java/com/example/musing/entity/music/PlayList.java
@@ -1,0 +1,25 @@
+package com.example.musing.entity.music;
+
+import com.example.musing.entity.user.User;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter // Lombok 어노테이션 : 클래스 내 모든 필드의 Getter 메소드 자동 생성
+@NoArgsConstructor // Lombok 어노테이션 : 기본 생성자 자동 추가
+@Entity
+@Table(name="playlist")
+public class PlayList {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column(nullable = false)
+    private String listname;
+
+    //유저에 관한 외래키 보유 주인테이블
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user; // 작성자
+}
