+++ src/main/java/com/example/musing/entity/user/Auth.java
@@ -1,7 +1,11 @@
-package com.example.musing.Entity;
+package com.example.musing.entity.user;
 
 import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@Getter // Lombok 어노테이션 : 클래스 내 모든 필드의 Getter 메소드 자동 생성
+@NoArgsConstructor // Lombok 어노테이션 : 기본 생성자 자동 추가
 @Entity
 @Table(name="auth")
 public class Auth {
