+++ src/main/java/com/example/musing/entity/user/User.java
@@ -1,0 +1,56 @@
+package com.example.musing.entity.user;
+
+import com.example.musing.entity.Board.Board;
+import com.example.musing.entity.Board.Reply;
+import com.example.musing.entity.music.PlayList;
+import com.example.musing.entity.music.Prefer;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Getter // Lombok 어노테이션 : 클래스 내 모든 필드의 Getter 메소드 자동 생성
+@NoArgsConstructor // Lombok 어노테이션 : 기본 생성자 자동 추가
+@Entity
+@Table(name = "user")
+public class User {
+
+
+    @Id
+    @GeneratedValue
+    @Column(name="userid")// AUTO_INCREMENT
+    private Long id;
+
+    @Column(nullable = false, length = 50)
+    private String username; // 사용자 이름
+
+    @Column(nullable = false, unique = true, length = 100)
+    private String email; // 이메일 (유니크)
+
+    @Column(nullable = false)
+    private String password;
+
+    // true면 사용가능 false면 정지상태
+    @Column(nullable = false)
+    private boolean Activated;
+
+
+
+
+    //게시판 일대다 매핑
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Board> boards = new ArrayList<>();
+    //플레이 리스트 일대다 매핑
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<PlayList> playLists = new ArrayList<>();
+    //플레이 리스트 일대다 매핑
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Prefer> prefers = new ArrayList<>();
+    //답글 리스트와 일대다 매핑
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Reply> replyList = new ArrayList<>();
+
+
+}
