+++ src/main/java/com/example/musing/repository/music/MusicRepository.java
@@ -1,6 +1,6 @@
 
-package com.example.musing.Repository;
-import com.example.musing.Entity.Music;
+package com.example.musing.repository.music;
+import com.example.musing.entity.music.Music;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MusicRepository extends JpaRepository<Music, Long> {
