+++ src/main/java/com/example/musing/repository/music/PlayListRepository.java
@@ -1,6 +1,6 @@
 
-package com.example.musing.Repository;
-import com.example.musing.Entity.PlayList;
+package com.example.musing.repository.music;
+import com.example.musing.entity.music.PlayList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlayListRepository extends JpaRepository<PlayList, Long> {
