+++ src/main/java/com/example/musing/repository/music/PreferRepository.java
@@ -1,6 +1,6 @@
 
-package com.example.musing.Repository;
-import com.example.musing.Entity.Prefer;
+package com.example.musing.repository.music;
+import com.example.musing.entity.music.Prefer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PreferRepository extends JpaRepository<Prefer, Long> {
