+++ src/main/java/com/example/musing/repository/user/UserRepository.java
@@ -1,6 +1,6 @@
 
-package com.example.musing.Repository;
-import com.example.musing.Entity.User;
+package com.example.musing.repository.user;
+import com.example.musing.entity.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
